+++ web/app/themes/brocooly/functions.php
@@ -21,18 +21,6 @@ require_once __DIR__ . '/bootstrap/app.php';
 
 /**
  * -------------------------------------------------------------------------
- * Include Kirki Framework
- * -------------------------------------------------------------------------
- *
- * If your theme requires theme options
- * you may include this file
- */
-if ( ! class_exists( \Kirki\Compatibility\Kirki::class ) ) {
-	require_once WP_CONTENT_DIR . '/vendor/aristath/kirki/kirki.php';
-}
-
-/**
- * -------------------------------------------------------------------------
  * Disable WordPress core update notification
  * -------------------------------------------------------------------------
  *